requirements.txt
APScheduler==3.10.4
firebase-admin==6.3.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
uuid0==0.2.7